requirements.txt
aiogram==2.21
aiohttp
aioredis==2.0.1
asyncpg==0.25.0
gunicorn
python-dotenv==0.20.0
sqlalchemy==1.4.39